Complete basic plotting script
This plotting script provides the bare minimum diagnostic plotting
required for the instrument display screen to be useful.

The following information is required:

- Which statistic should be reported per sensor? Currently, 3 sensor
median is being plotted
- Are there any scaling factors that need to be applied in order to
convert different values into PPM?
- Can this code be easily run on the Raspberry Pi using the distribution
Python? Do I need to compile and install the additional libraries
(pyqtgraph) in order for it to work?
- Do I need to do anything to the graphics to make them more useful in a
demonstration? Plot spread, perhaps?

The following needs to be done on the Pi:

- Set the plotting code to run after a specified time when the system
starts. See sensor_array_utils for a method, if needs be