Update control scripts for WACL CO2 experiment (#2)
Update various names to match WACL CO2 experiment - a lot of this stuff
should really be abstracted into a configuration file. That will be
straightforward to do when I have time.

A minor update is the inclusion of timestamps in the control log, which
should make remotely debugging issues more straightforward.