Improve data uploading procedure
Instead of looking for the second to last file in the sorted data
directory, look instead for a log file containing yesterday's date in
the name. If it doesn't exist, or there is more than one (?!), don't do
anything further - the lack of an upload will trigger the monitoring
app and notify project collaborators that something is wrong.