Removed archive's arguments
The data log directory is now specified in `config.py` so it doesn't
need to be passed in to the archiving script.
Previously, the data log directory was hardcoded in `data_writer` but
passed in at runtime to `archive`, now it is hardcoded for both at Deb
build time. An argument could be made that it should be specified at
runtime for both.