Working with Bullseye release
The latest version of RaspberryPi OS is based off Debian Bullseye and
has necessitated a few minor changes to the package.

The most significant is that the video drivers have been changed and the
new ones do not work with the touchscreen display used in the BOCS. The
old drivers can be selected in `raspi-config` - this change has been
reflected in the `README`.

The other change is that the udev rules take a little time to propagate
to allow the Pi to see the new serial connection. This just requires
adding a brief pause before starting the data collection.

Other changes in this commit are minor housekeeping of build
dependencies.