Refactored data_writer
The logic flow in write_data has been slightly changed for several
reasons. Firstly so that when there is an error in the received data
the usual logging doesn't proceed. Secondly to place error handling
closer to the code it is raised from, and finally to provide more detail in
the logs themselves.

The logic in line 52 hasn't been changed, although I don't understand
in what kind of scenario this might arise or be useful.

There are multiple return points which isn't ideal but since the first
occurs when an error is encountered it isn't unexpected